requirements.txt
dash
plotly
numpy
requests
google-generativeai